@@ -1,53 +1,102 @@
 {
   "name": "vue-dual-list",
-  "version": "0.0.7",
-  "description": "",
+  "version": "1.0.0",
+  "description": "Dual List in VueJs with VueMaterial",
+  "author": "mikemajesty <mikemajesty2011@gmail.com>",
   "main": "component/vue-dual-list.vue",
   "scripts": {
-    "compile": "webpack --config ./webpack/webpack.config.js --inline --hot"
+    "dev": "nodemon build/dev-server.js",
+    "start": "nodemon build/dev-server.js",
+    "build": "nodemon build/build.js",
+    "unit": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
+    "e2e": "nodemon test/e2e/runner.js",
+    "test": "npm run unit && npm run e2e",
+    "lint": "eslint --ext .js,.vue demo test/unit/specs test/e2e/specs",
+    "compile": "webpack --config ./webpack/webpack.config.js --hot"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/mikemajesty/Vue-Dual-List.git"
-  },
-  "keywords": [
-    "vue",
-    "webpack",
-    "loader"
-  ],
-  "author": "Mike Lima",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mikemajesty/Vue-Dual-List/issues"
-  },
-  "homepage": "https://github.com/mikemajesty/Vue-Dual-List#readme",
   "dependencies": {
-    "babel-loader": "6.4.1",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "express": "4.15.2",
+    "vue": "^2.2.6",
+    "vue-router": "^2.3.1",
+    "vue-material": "^0.7.1",
     "serve-favicon": "2.3.2",
-    "vue": "2.2.6",
-    "vue-loader": "11.3.4",
-    "vue-material": "0.7.1",
-    "vue-template-compiler": "^2.2.6",
-    "webpack-dev-server": "1.16.3"
+    "express": "4.15.2"
   },
   "devDependencies": {
-    "babel-helper-vue-jsx-merge-props": "2.0.2",
-    "babel-plugin-syntax-jsx": "6.18.0",
-    "babel-plugin-transform-vue-jsx": "3.4.2",
-    "babel-preset-es2015": "6.24.0",
-    "babel-preset-stage-2": "6.22.0",
-    "eslint": "3.13.1",
-    "gulp": "3.9.1",
-    "gulp-babel": "6.1.2",
-    "gulp-concat": "2.6.0",
-    "gulp-uglify": "1.5.3",
-    "gulp-uglifycss": "1.0.6",
-    "vueify": "9.4.1",
-    "webpack": "1.12.9",
-    "webpack-combine-loaders": "2.0.0",
-    "webpack-dev-server": "1.14.0",
-    "css-loader": "0.25.0"
-  }
+    "autoprefixer": "^6.7.2",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^7.1.1",
+    "babel-helper-vue-jsx-merge-props": "^2.0.2",
+    "babel-loader": "^6.2.10",
+    "babel-plugin-istanbul": "^4.1.1",
+    "babel-plugin-syntax-jsx": "^6.18.0",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-plugin-transform-vue-jsx": "^3.4.2",
+    "babel-preset-env": "^1.3.2",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-register": "^6.22.0",
+    "chai": "^3.5.0",
+    "chalk": "^1.1.3",
+    "chromedriver": "^2.27.2",
+    "connect-history-api-fallback": "^1.3.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "cross-env": "^4.0.0",
+    "cross-spawn": "^5.0.1",
+    "css-loader": "^0.28.0",
+    "eslint": "^3.19.0",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-friendly-formatter": "^2.0.7",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-html": "^2.0.0",
+    "eslint-plugin-promise": "^3.4.0",
+    "eslint-plugin-standard": "^2.0.1",
+    "eventsource-polyfill": "^0.9.6",
+    "express": "^4.14.1",
+    "extract-text-webpack-plugin": "^2.0.0",
+    "file-loader": "^0.11.1",
+    "friendly-errors-webpack-plugin": "^1.1.3",
+    "html-webpack-plugin": "^2.28.0",
+    "http-proxy-middleware": "^0.17.3",
+    "inject-loader": "^3.0.0",
+    "karma": "^1.4.1",
+    "karma-coverage": "^1.1.1",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-phantomjs-shim": "^1.4.0",
+    "karma-sinon-chai": "^1.3.1",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-spec-reporter": "0.0.30",
+    "karma-webpack": "^2.0.2",
+    "lolex": "^1.5.2",
+    "mocha": "^3.2.0",
+    "nightwatch": "^0.9.12",
+    "opn": "^4.0.2",
+    "optimize-css-assets-webpack-plugin": "^1.3.0",
+    "ora": "^1.2.0",
+    "phantomjs-prebuilt": "^2.1.14",
+    "rimraf": "^2.6.0",
+    "selenium-server": "^3.0.1",
+    "semver": "^5.3.0",
+    "shelljs": "^0.7.6",
+    "sinon": "^2.1.0",
+    "sinon-chai": "^2.8.0",
+    "url-loader": "^0.5.8",
+    "vue-loader": "^11.3.4",
+    "vue-style-loader": "^2.0.5",
+    "vue-template-compiler": "^2.2.6",
+    "webpack": "^2.3.3",
+    "webpack-bundle-analyzer": "^2.2.1",
+    "webpack-dev-middleware": "^1.10.0",
+    "webpack-hot-middleware": "^2.18.0",
+    "webpack-merge": "^4.1.0"
+  },
+  "engines": {
+    "node": ">= 4.0.0",
+    "npm": ">= 3.0.0"
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not ie <= 8"
+  ]
 }
